refactor: Product 가격 검증책임 도메인으로 이동
