+++ src/main/java/kitchenpos/domain/Product.java
@@ -1,6 +1,7 @@
 package kitchenpos.domain;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 
 public class Product {
     private final Long id;
@@ -8,9 +9,16 @@ public class Product {
     private final BigDecimal price;
 
     public Product(final Long id, final String name, final BigDecimal price) {
+        validate(price);
         this.id = id;
         this.name = name;
         this.price = price;
+    }
+
+    private void validate(final BigDecimal price) {
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("상품의 가격이 음수가 될 수 없습니다.");
+        }
     }
 
     public Product(final String name, final BigDecimal price) {
